[REF] TransactionTable: drop for show
The transaction table was being too complicated and more
page related. This commit removes it to render on transactions
show page, as an Outlet is needed into it.
