+++ app/components/Transaction/index.ts
@@ -1,4 +1,2 @@
 export { default as TransactionCard } from './TransactionCard';
 export type { TransactionCardProps } from './TransactionCard';
-export { default as TransactionTable } from './TransactionTable';
-export type { TransactionTableProps } from './TransactionTable';
