+++ app/services/transactions/RawTransaction.ts
@@ -1,0 +1,9 @@
+export interface RawTransaction {
+    description: string;
+    occurred_at: string;
+    value: string;
+    currency: string;
+    category_name: string;
+    category_icon: number;
+    uuid: string;
+}
